update : Generalize DeleteDialog components for better usability
